+++ resources/views/layouts/app.blade.php
@@ -71,7 +71,7 @@
         })();
     </script>
 
-    <link rel="stylesheet" href="{{ asset('vendor/isproject/isproject.css') }}">
+    <link rel="stylesheet" href="{{ isproject_asset('isproject.css') }}">
 
     @include('isproject::partials.accent')
 
@@ -100,15 +100,15 @@
     </div>
 </div>
 
-<script src="{{ asset('vendor/isproject/bootstrap.bundle.min.js') }}"></script>
+<script src="{{ isproject_asset('bootstrap.bundle.min.js') }}"></script>
 
 {{-- Searchable dropdowns. Before ours, which looks for TomSelect and simply
      leaves the selects alone if it is not there. --}}
 @if (config('isproject.select.enabled', true))
-    <script src="{{ asset('vendor/isproject/tom-select.min.js') }}"></script>
+    <script src="{{ isproject_asset('tom-select.min.js') }}"></script>
 @endif
 
-<script src="{{ asset('vendor/isproject/isproject.js') }}"></script>
+<script src="{{ isproject_asset('isproject.js') }}"></script>
 @stack('scripts')
 </body>
 </html>